Removed old code, renamed new code files
