--- CodeGenerator/Sources/CodeGenerator/Doc/Paginator.swift
@@ -1,21 +1,0 @@
-//===----------------------------------------------------------------------===//
-//
-// This source file is part of the AWSSDKSwift open source project
-//
-// Copyright (c) 2017-2020 the AWSSDKSwift project authors
-// Licensed under Apache License v2.0
-//
-// See LICENSE.txt for license information
-// See CONTRIBUTORS.txt for the list of AWSSDKSwift project authors
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-//===----------------------------------------------------------------------===//
-
-struct Paginator {
-    let methodName: String
-    let resultKeys: [String]
-    let inputTokens: [String]
-    let outputTokens: [String]
-    let limitKey: String
-}
